# docs/swagger.yaml
definitions:
...
  github_com_evrone_go-clean-template_internal_entity.AISettings:
    properties:
      includeGuessedCorrect:
        type: boolean
      revisionEnabled:
        type: boolean
      revisionIntervalsDays:
        items:
          type: integer
        type: array
      strongThresholdPercent:
        type: integer
      weaknessMinAttempts:
        type: integer
      weaknessThresholdPercent:
        type: integer
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.AdminEventSummary:
    properties:
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      id:
        type: string
      name:
        type: string
      registeredCount:
        type: integer
      startAt:
        type: string
      status:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamStatus'
      type:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamConfigType'
    type: object
  github_com_evrone_go-clean-template_internal_entity.AdminEventsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminEventSummary'
        type: array
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.AnalyticsOverview:
    properties:
      activeUsers:
        type: integer
      averageAccuracy:
        type: number
      averageStudyMinutes:
        type: number
      questionsAnswered:
        type: integer
      totalRewards:
        type: integer
      totalUsers:
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_entity.AnalyticsPoint:
    properties:
      date:
        type: string
      value:
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_entity.AnalyticsTimeSeries:
    properties:
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      metric:
        type: string
      points:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AnalyticsPoint'
        type: array
      range:
        type: string
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.Coupon:
    properties:
      amount:
        type: integer
      code:
        type: string
      description:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      maxUsesPerUser:
        type: integer
      maxUsesTotal:
        type: integer
      type:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.CouponCreateRequest:
    properties:
      amount:
        type: integer
      code:
        type: string
      description:
        type: string
      expiresAt:
        type: string
      isActive:
        type: boolean
      maxUsesPerUser:
        type: integer
      maxUsesTotal:
        type: integer
      type:
        type: string
    required:
    - code
    - type
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.ExamCategory:
    enum:
    - NEET_PG
    - NEET_UG
    - JEE
    - UPSC
    type: string
    x-enum-varnames:
    - ExamCategoryNEETPG
    - ExamCategoryNEETUG
    - ExamCategoryJEE
    - ExamCategoryUPSC
...
  github_com_evrone_go-clean-template_internal_entity.ExamConfigType:
    enum:
    - MOCK
    - SUBJECT_TEST
    - REWARD_EVENT
    - DAILY_TEST
    type: string
    x-enum-varnames:
    - ExamTypeMock
    - ExamTypeSubjectTest
    - ExamTypeRewardEvent
    - ExamTypeDailyTest
...
  github_com_evrone_go-clean-template_internal_entity.ExamStatus:
    enum:
    - DRAFT
    - SCHEDULED
    - ONGOING
    - COMPLETED
    type: string
    x-enum-varnames:
    - ExamStatusDraft
    - ExamStatusScheduled
    - ExamStatusOngoing
    - ExamStatusCompleted
...
  github_com_evrone_go-clean-template_internal_entity.SubjectAccuracyItem:
    properties:
      accuracy:
        type: number
      subjectId:
        type: string
      subjectName:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.SubjectAccuracyResponse:
    properties:
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      subjects:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.SubjectAccuracyItem'
        type: array
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.WeakTopicItem:
    properties:
      accuracy:
        type: number
      attempts:
        type: integer
      subjectId:
        type: string
      subjectName:
        type: string
      topicId:
        type: string
      topicName:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.WeakTopicsResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.WeakTopicItem'
        type: array
    type: object
  internal_controller_http_v1.ErrorResponse:
    properties:
      error:
        example: message
        type: string
    type: object
info:
  contact: {}
paths:
  /admin/ai-settings:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AISettings'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Get AI settings
      tags:
      - 'Admin: AI Settings'
    put:
      consumes:
      - application/json
      parameters:
      - description: Settings payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AISettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AISettings'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Update AI settings
      tags:
      - 'Admin: AI Settings'
  /admin/analytics/overview:
    get:
      parameters:
      - description: Exam
        in: query
        name: exam
        type: string
      - description: today,7d,30d
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AnalyticsOverview'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Analytics overview
      tags:
      - 'Admin: Analytics'
  /admin/analytics/subject-accuracy:
    get:
      parameters:
      - description: Exam
        in: query
        name: exam
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.SubjectAccuracyResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Subject accuracy
      tags:
      - 'Admin: Analytics'
  /admin/analytics/time-series:
    get:
      parameters:
      - description: Metric (active_users|questions_answered)
        in: query
        name: metric
        required: true
        type: string
      - description: Exam
        in: query
        name: exam
        type: string
      - description: today|7d|30d
        in: query
        name: range
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AnalyticsTimeSeries'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Analytics time series
      tags:
      - 'Admin: Analytics'
  /admin/analytics/weak-topics:
    get:
      parameters:
      - description: Exam
        in: query
        name: exam
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.WeakTopicsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Weak topics
      tags:
      - 'Admin: Analytics'
  /admin/coupons:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Coupon'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: List coupons
      tags:
      - 'Admin: Coupons'
    post:
      consumes:
      - application/json
      parameters:
      - description: Coupon payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.CouponCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Coupon'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Create coupon
      tags:
      - 'Admin: Coupons'
  /admin/coupons/{id}:
    delete:
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Delete coupon
      tags:
      - 'Admin: Coupons'
    get:
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Coupon'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Get coupon
      tags:
      - 'Admin: Coupons'
    patch:
      consumes:
      - application/json
      parameters:
      - description: Coupon ID
        in: path
        name: id
        required: true
        type: string
      - description: Coupon payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.CouponCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Coupon'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Update coupon
      tags:
      - 'Admin: Coupons'
  /admin/events/upcoming:
    get:
      parameters:
      - description: Exam
        in: query
        name: exam
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.AdminEventsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Upcoming events
      tags:
      - 'Admin: Events'
...
swagger: "2.0"