# internal/controller/http/v1/doc.go
package v1

import (
	responsepkg "github.com/evrone/go-clean-template/internal/controller/http/v1/response"
)

// ErrorResponse is re-exported for Swagger annotations.
type ErrorResponse = responsepkg.Error